Fixed invalid use of dublicate context usage.
The problem was that in main.c the I2C hi2cl context is used to initalize
the IC2 port and the hi2cl structure itself. Another context was used
to read from the device but this context is never initialized.

The current implementation should not have worked, but it did...