Added CA protocol layer
This layer should replace the HandleGeneric message layer. This "message layer"
is a sort of CA module protocol and should be handled as a protocol layer, i.e
data->layer->application.

This input (data) is sent to the layer via inputCAProtocol and application handlers
is defined in the struct CAProtocolCtx which is initialized via initCAProtocol.

This layer is introduced to make communication generic and prepare for common
calibration.