Issue #78, Added HW watch dog to AC, DC and Temperature board
The hardware watchdog is enabled/working in this commit but once enabled
it can not be disabled. From reference manual

Enabling the watchdog
The watchdog is always disabled after a reset. It is enabled by setting
the WDGA bit in the WWDG_CR register, then it cannot be disabled again
except by a reset.

Unfortuantly the STM DFU SW uploader code does not handle/polls/kicks
the hwdg which means that HW Watch dog must not be active during DFU SW
upload, else the hwdg is triggered.

The watch dog is required and a new procedure has been created for SW update
1 - User enter DFU<enter>
2 - Board makes a SW reset
3 - On boot, it is detected that a SW reset is made and system enter a
    SW update routing. Board sends "DFU mode: press y<enter> to start SW update"
    a - user enter y<enter> and SW update via DFU is active
    b - user does nothing and system performs a HW Watch Dog reset after 30 seconds.

Further, this commit also changes the message to user on reconnect.
New message "reconnected Reset Reason: <reset reason>"

The functions has been made generic and moved to CAProtocolStm.