Made USB Rx/Tx handling more robust and simple
The Circular buffer should only handle bytes, not anything about
protocol upper layer issues. It is used between lover/upper half
of an IRQ handler and should be as simple as possible. The buffer
containing bytes has been moved inside the opaque struct cbuf_handle_t,
since it has no purpose outside of the curcular buffer.

The CA protocol handler has been changed so that bytes is fetched from
the circular buffer and moved to a new buffer in an opaque struct in
CA_proto. When 'r' or 'n' is received a message is detected and handed to
the protocol parser, inputCAProtocol. This makes the preveous message
counter in circular buffer obsolete. Protocol issues should be handled
in CA_proto, not in an IRQ handler.